+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Limudim support for Parsha
+- Kviah function for JewishDate
+
 ### Fixed
 - Support for python-dateutil pre-2.7.0